removing exported functions from routers
